requirements.txt
Flask
streamlit
pillow
pymupdf
PyPDF2
